requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
pymongo==4.6.0
chromadb==0.4.18
google-generativeai==0.3.1
python-dotenv==1.0.0
python-multipart==0.0.6